Initial Commit and bringing the solution to run. Adding big-instance-explorer project. Most of these will be reverted before final commit.

+++ client/esbuild.config.mts
@@ -50,7 +50,20 @@ export class ESBuildRunner {
             throw new Error('No outdir specified');
         }
 
-        fs.rmSync(this.config.outdir, { recursive: true, force: true });
+        if (this.isWatch) {
+            return;
+        }
+
+        try {
+            fs.rmSync(this.config.outdir, { recursive: true, force: true });
+        } catch (error) {
+            const code = (error as NodeJS.ErrnoException).code;
+            if (code !== 'EPERM' && code !== 'EACCES') {
+                throw error;
+            }
+
+            console.warn(`Skipping cleanup for '${this.config.outdir}' because it is locked by another process.`);
+        }
     }
 
     async run(): Promise<void> {
